начало разбивки и импорта
