+++ index.html
@@ -227,7 +227,7 @@
     </section>
   </template>
   <script src="js/functions.js" defer></script>
-  <script src="js/main.js" defer></script>
+  <script src="main.js" type="module"></script>
 
 </body>
 </html>
